requirements.txt
streamlit
websocket-client
pandas
plotly
pyserial 